refactor(media): 搜索归一化上提到媒体层，游戏侧改委托（P5-A 地基）
书架与视频库页此前**没有搜索**，只有游戏库页有一套日语感知的归一化实现
（全角转半角 / 片假名转平假名 / 丢标点与长音符）。补搜索时不该各写一套
「大概也做全角转半角」的近似实现——同一个库、同一批日文标题，在三个页面
搜出不同结果比没有搜索更难解释。

- 新增 `hibiki/lib/src/media/media_search_text.dart`：`normalizeMediaSearchText`
  （实现原样上提，逐字符逻辑未改）+ `matchesMediaSearch`（归一化子串匹配，
  空查询恒命中，多标题任一命中即算命中）。
- `galgame_library_query.dart` 的 `normalizeGalgameSearchText` /
  `matchesGalgameSearch` 改为委托，删掉重复的丢弃字符集副本。**保留原函数名**：
  它们是该文件公开 API，已被游戏库页与单测引用，改名无收益只破坏调用方。

本次是纯重构，无行为变化：test/mining 799 passed（含游戏搜索归一化/匹配全部
既有用例）；analyze clean。书架/视频页的搜索 UI 接入在后续 commit。
